free_shopping_chatbot/backend/requirements.txt
# Dependencias del backend
fastapi
uvicorn
google-genai